requirements.txt
django==4.2.10
psycopg2==2.9.7
djangorestframework
django-cors-headers
djangorestframework-simplejwt